+++ changelog.md
@@ -4,6 +4,7 @@
 ### Version 0.9.0
 
 - Introduced `System::make_group_whole` for making a group "whole" in the system (similar to `System::make_molecules_whole`).
+- **Breaking change:** `System::group_all_distances` now returns ndarray's `Array2` instead of a vector of vectors.
 
 #### Bug fixes and minor changes
 - `System::traj_iter_map_reduce` now properly propagates errors and failures from the individual threads into the master-thread `ProgressPrinter`.